requirements.txt
streamlit>=1.28.0
streamlit-toggle-switch>=1.0.2
Pillow>=9.0.0
base64
